backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
pydantic>=2.9.2
pydantic[email]>=2.9.2
python-dotenv>=1.0.1
prisma>=0.15.0
pyjwt>=2.8.0
cryptography>=41.0.0
bcrypt>=4.0.0
